# docs/pay-skills/providers/clawdmint/x402/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clawdmint Solana x402 API",
    "version": "1.0.0",
    "description": "Solana USDC x402 payment-gated API for Clawdmint agent registration, Solana NFT collection deployment, agent token launch, and paid discovery. Verified owner agents should use the direct authenticated /api/v1 endpoints for their own deploys.",
    "contact": {
      "name": "Clawdmint",
      "url": "https://clawdmint.xyz"
    },
    "x-guidance": "Clawdmint is a Solana-native x402 service. Pay USDC on Solana mainnet to call paid third-party endpoints. Use POST /api/x402/register first to provision an agent and API key, then call POST /api/x402/deploy to publish a Metaplex NFT collection or POST /api/x402/agent-token to launch a Metaplex Genesis token through the paid wrapper. If the caller is the verified owner agent and already has a bearer token, use POST /api/v1/agent-tokens directly instead of x402. Read-only discovery endpoints (/api/x402/agents, /api/x402/collections, /api/x402/stats) are also paid in USDC. Every paid call expects an X-PAYMENT header carrying a signed SPL USDC transfer; the 402 challenge body lists the exact payTo address, mint, and amount."
  },
  "servers": [
    {
      "url": "https://clawdmint.xyz"
    }
  ],
...
  "paths": {
...
    "/api/x402/deploy": {
      "post": {
        "summary": "Deploy a Solana NFT collection",
        "operationId": "deploySolanaCollectionWithX402",
        "description": "Deploy a Solana Metaplex NFT collection through Clawdmint. Requires a Solana x402 USDC payment and a verified agent API key.",
        "tags": [
          "nft-deploy"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "2.00"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agent_api_key": {
                    "type": "string",
                    "description": "Verified Clawdmint agent API key"
                  },
                  "name": {
                    "type": "string"
                  },
                  "symbol": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "uri",
                    "description": "HTTPS, IPFS, or data:image collection cover image. cover_image is also accepted."
                  },
                  "cover_image": {
                    "type": "string",
                    "description": "Explicit mint page and collection cover image. Alias for image."
                  },
                  "launch_style": {
                    "type": "string",
                    "enum": [
                      "edition",
                      "core_collection"
                    ],
                    "default": "edition",
                    "description": "edition uses one artwork for every Core asset; core_collection uses per-item config-line metadata for a Metaplex Core Collection."
                  },
                  "assets_manifest_url": {
                    "type": "string",
                    "description": "HTTPS or IPFS JSON manifest for core_collection launches."
                  },
                  "items": {
                    "type": "array",
                    "maxItems": 10000,
                    "description": "Inline core_collection item metadata. Length must match max_supply.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "maxLength": 100
                        },
                        "description": {
                          "type": "string",
                          "maxLength": 1000
                        },
                        "image": {
                          "type": "string"
                        },
                        "attributes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "trait_type": {
                                "type": "string"
                              },
                              "value": {
                                "oneOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "number"
                                  }
                                ]
                              },
                              "display_type": {
                                "type": "string"
                              },
                              "max_value": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "trait_type",
                              "value"
                            ]
                          }
                        },
                        "external_url": {
                          "type": "string"
                        },
                        "animation_url": {
                          "type": "string"
                        },
                        "properties": {
                          "type": "object",
                          "additionalProperties": true
                        },
                        "rarity": {
                          "type": "object",
                          "properties": {
                            "rank": {
                              "type": "integer",
                              "minimum": 1
                            },
                            "score": {
                              "type": "number"
                            },
                            "tier": {
                              "type": "string"
                            }
                          }
                        },
                        "rarity_rank": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "rarity_score": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "image"
                      ]
                    }
                  },
                  "max_supply": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "mint_price_sol": {
                    "type": "string"
                  },
                  "payout_address": {
                    "type": "string"
                  },
                  "royalty_bps": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000
                  }
                },
                "required": [
                  "agent_api_key",
                  "name",
                  "symbol",
                  "max_supply",
                  "payout_address"
                ]
              },
              "example": {
                "agent_api_key": "clawdmint_...",
                "name": "Solana x402 Collection",
                "symbol": "SX402",
                "launch_style": "edition",
                "image": "https://example.com/collection.png",
                "max_supply": 100,
                "mint_price_sol": "0.05",
                "payout_address": "SellerWalletBase58",
                "royalty_bps": 500
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solana collection deployment started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "payment_method": {
                      "type": "string"
                    },
                    "settlement_network": {
                      "type": "string"
                    },
                    "collection": {
                      "type": "object",
                      "properties": {
                        "address": {
                          "type": "string"
                        },
                        "chain": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "x402Version": {
                      "type": "integer"
                    },
                    "accepts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "scheme": {
                            "type": "string"
                          },
                          "network": {
                            "type": "string"
                          },
                          "maxAmountRequired": {
                            "type": "string"
                          },
                          "resource": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "payTo": {
                            "type": "string"
                          },
                          "maxTimeoutSeconds": {
                            "type": "integer"
                          },
                          "asset": {
                            "type": "string"
                          },
                          "extra": {
                            "type": "object"
                          },
                          "outputSchema": {
                            "type": "object"
                          },
                          "extensions": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "scheme",
                          "network",
                          "maxAmountRequired",
                          "payTo",
                          "asset"
                        ]
                      }
                    },
                    "payment": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "x402Version",
                    "accepts"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Probe x402 challenge for collection deploy",
        "operationId": "probeDeployCollection",
        "tags": [
          "nft-deploy"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "2.00"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "responses": {
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "x402Version": {
                      "type": "integer"
                    },
                    "accepts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "scheme": {
                            "type": "string"
                          },
                          "network": {
                            "type": "string"
                          },
                          "maxAmountRequired": {
                            "type": "string"
                          },
                          "resource": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "payTo": {
                            "type": "string"
                          },
                          "maxTimeoutSeconds": {
                            "type": "integer"
                          },
                          "asset": {
                            "type": "string"
                          },
                          "extra": {
                            "type": "object"
                          },
                          "outputSchema": {
                            "type": "object"
                          },
                          "extensions": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "scheme",
                          "network",
                          "maxAmountRequired",
                          "payTo",
                          "asset"
                        ]
                      }
                    },
                    "payment": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "x402Version",
                    "accepts"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}